# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API server .",
        "title": "金流SERVER API",
        "contact": {
            "name": "CQI-SERVICES",
            "url": "https://www.cqiserv.com/",
            "email": "zor@cqiserv.com"
        },
        "license": {},
        "version": "2020.08"
    },
    "paths": {
        "/admin/order/{OrderSubID}": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Order"
                ],
                "summary": "OrderGet 查詢交易單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易單子單號",
                        "name": "OrderSubID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        },
        "/admin/orders": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Order"
                ],
                "summary": "OrderGetAll 查詢ALL帳號",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Order"
                            }
                        }
                    }
                }
            }
        },
        "/admin/platform": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Platform"
                ],
                "summary": "PlatformRegisterServer 新增平台帳號",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台名稱",
                        "name": "PlatformName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台帳號",
                        "name": "PlatformAccount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台密碼",
                        "name": "PlatformPassword",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "聯絡信箱",
                        "name": "PlatformEmail",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "群組代號 - 預設為1",
                        "name": "PlatformGroupID",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        },
        "/admin/platform/status/{PlatformAccount}": {
            "put": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Platform"
                ],
                "summary": "PlatformStatusEnable 啟用帳號",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台帳號",
                        "name": "PlatformAccount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Platform"
                ],
                "summary": "PlatformStatusDisable 停用帳號",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台帳號",
                        "name": "PlatformAccount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        },
        "/admin/platform/{PlatformAccount}": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Platform"
                ],
                "summary": "PlatformGet 查詢帳號",
                "parameters": [
                    {
                        "type": "string",
                        "description": "平台帳號",
                        "name": "PlatformAccount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        },
        "/admin/platforms": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Platform"
                ],
                "summary": "PlatformGetAll 查詢ALL帳號",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Platform"
                            }
                        }
                    }
                }
            }
        },
        "/casino/shop/": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Casino"
                ],
                "summary": "GetItem 查詢casino商品清單",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/casinogrpc.CasinoItem"
                        }
                    }
                }
            }
        },
        "/casino/user/{UserID}": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Casino"
                ],
                "summary": "GetItem 查詢casino用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CASINO角色帳號",
                        "name": "UserID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        },
        "/mycardsandbox/order": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MyCard-SandBox"
                ],
                "summary": "新增 mycard 交易單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家帳號ID",
                        "name": "OrderClientID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品代號",
                        "name": "OrderItemID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品價格",
                        "name": "OrderItemPrice",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台token代號",
                        "name": "PlatformToken",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    },
                    "411": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "casinogrpc.CasinoItem": {
            "type": "object",
            "properties": {
                "itemName": {
                    "type": "string"
                },
                "itemType": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "productID": {
                    "type": "string"
                }
            }
        },
        "controller.Resp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "ecode": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "db.Order": {
            "type": "object",
            "properties": {
                "callbackOriginalData": {
                    "type": "string"
                },
                "callbackURL": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderClientID": {
                    "type": "string"
                },
                "orderDate": {
                    "type": "string"
                },
                "orderGameSubID": {
                    "type": "string"
                },
                "orderItemID": {
                    "type": "string"
                },
                "orderItemPrice": {
                    "type": "string"
                },
                "orderOriginalData": {
                    "type": "string"
                },
                "orderSubID": {
                    "type": "string"
                },
                "paymentID": {
                    "type": "string"
                },
                "paymentTypeID": {
                    "type": "integer"
                },
                "platformID": {
                    "type": "integer"
                },
                "receivedCallbackDate": {
                    "type": "string"
                },
                "stageType": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.Platform": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "platformAccount": {
                    "type": "string"
                },
                "platformEmail": {
                    "type": "string"
                },
                "platformGroupID": {
                    "type": "integer"
                },
                "platformName": {
                    "type": "string"
                },
                "platformPassword": {
                    "type": "string"
                },
                "platformToken": {
                    "type": "string"
                },
                "platformTokenSecret": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}